Error check around sell unit with no position given and try catch around messages. - Ninja

+++ src/gachaRoutes.js
@@ -230,18 +230,26 @@ async function sellUnit(message) {
 
   var args = getArgs(message);
   try{
-    var unitToSell = args[0];
-    let currentUser = await user.getUser(message.author.id);
-    let sellUnit = await units.getUnit(message.author.id, unitToSell);
 
-    let confirmText = message.guild.member(message.author).displayName + " are you sure you want to sell "
-    + "Lv" + Math.floor(sellUnit.lvl) + " " + iconsunit.getRankIcon(sellUnit.rank) + " " + sellUnit.unit_name + " for "+ sellUnit.getCloverValue() + iconscurr.getCurrencyIcon(forCurrency) + "?"
+    if (args[0] !== undefined) {
+      var unitToSell = args[0];
+      let currentUser = await user.getUser(message.author.id);
+      let sellUnit = await units.getUnit(message.author.id, unitToSell);
 
-    if(await embeds.confirmationMessageYN(message, confirmText)){
-      let sellValue = sellUnit.getCloverValue();
-      if(await units.deleteUnit(sellUnit)){
-        cur.modCurrency(message, forCurrency, sellValue);
+      let confirmText = message.guild.member(message.author).displayName + " are you sure you want to sell "
+        + "Lv" + Math.floor(sellUnit.lvl) + " " + iconsunit.getRankIcon(sellUnit.rank) + " " + sellUnit.unit_name + " for " + sellUnit.getCloverValue() + iconscurr.getCurrencyIcon(forCurrency) + "?"
+
+      if (await embeds.confirmationMessageYN(message, confirmText)) {
+        let sellValue = sellUnit.getCloverValue();
+        if (await units.deleteUnit(sellUnit)) {
+          cur.modCurrency(message, forCurrency, sellValue);
+        }
       }
+    }
+    else
+    {
+      console.error("No unit position given");
+      embeds.printSingleError(message, "No unit position given");
     }
   }catch(err){
     console.error('Sell Unit Error : ' + err + " - " + err.stack);
